+++ contracts/Staking-Rental/StakedNetvrkTransport.sol
@@ -84,7 +84,7 @@ contract StakedNetvrkTransport is
         _ensureEOAorERC721Receiver(_stakeTo);
         require(_stakeTo != address(this), "INVALID_STAKE_TO");
         require(
-            _rentableUntil >=
+            uint256(_rentableUntil) >=
                 block.timestamp + (uint256(_minRentDays) * 1 days),
             "INVALID_RENTABLE_UNTIL"
         );
@@ -129,7 +129,7 @@ contract StakedNetvrkTransport is
         );
 
         require(
-            _rentableUntil >=
+            uint256(_rentableUntil) >=
                 block.timestamp + (uint256(_minRentDays) * 1 days),
             "INVALID_RENTABLE_UNTIL"
         );
